Create arduinohomecontrol.cpp
This project is an Arduino-based home automation application. It measures temperature and humidity using a DHT11 sensor, calculates the distance of nearby objects using an HC-SR04 ultrasonic distance sensor. It also adjusts the brightness of an LED with a potentiometer and produces a doorbell sound with a button press. The MFRC522 RFID module allows card reading capability, triggering specific actions when a particular card is detected. All data is displayed on an I2C LCD screen and also sent to the serial monitor.

This application can be utilized for indoor control and security purposes. For instance, it can be employed as a doorbell to notify of incoming visitors and facilitate access control through RFID cards.